Revert "PYTHON-4226 Address sign-compare warning"
This reverts commit 18cdb2634325eb81cc585cb07692437f8dc2c74c.

+++ bson/_cbsonmodule.c
@@ -2948,7 +2948,7 @@ static PyObject* _cbson_array_of_documents_to_buffer(PyObject* self, PyObject* a
             goto fail;
          }
 
-        if (view.len < (int32_t)size) {
+        if (view.len < size) {
             PyObject* InvalidBSON = _error("InvalidBSON");
             if (InvalidBSON) {
                 PyErr_SetString(InvalidBSON, "objsize too large");
